refactor(sync): extract sync/send.rs
Move the software-wallet send flow out of sync/mod.rs into a dedicated
send submodule:

  - propose_send / estimate_fee / execute_proposal (the 3-step
    propose → confirm → broadcast pipeline)
  - broadcast_raw_transaction (the SendTransaction gRPC helper)
  - ProposalResult (return type for propose_send)
  - zip317_helper (ZIP-317 change strategy + input selector factory,
    used by both propose_send and estimate_fee)
  - NoOpSpendProver / NoOpOutputProver (Sapling prover placeholders
    used for Orchard-only sends so the wallet doesn't have to ship
    the 50MB Sapling params)

The shared pieces stay in sync/mod.rs so the hardware PCZT pipeline
(sync/pczt.rs, extracted in the next commit) can also consume from
them without creating a cross-submodule cycle:

  - PROPOSAL_STORE (static LazyLock<Mutex<ProposalStore>>)
  - StoredProposal / ProposalStore (now pub(super) with pub fields)
  - open_wallet_db + WalletDatabase type alias (pub(super))

`propose_send`, `estimate_fee`, and `execute_proposal` are re-exported
from sync/mod.rs with `pub(crate) use send::{...};` so every existing
`crate::wallet::sync::propose_send` call path still resolves.

`ProposalResult` is NOT re-exported: api/sync.rs has its own
FRB-visible `ProposalResult` mirror and maps fields manually, so
nothing outside send.rs names the wallet-level type.

23 tests pass.
