+++ rust/src/wallet/sync/mod.rs
@@ -1,26 +1,15 @@
 use std::convert::Infallible;
-use std::num::NonZeroUsize;
 
 use rand::rngs::OsRng;
-use secrecy::{ExposeSecret, SecretVec};
 use zcash_client_backend::{
     data_api::{
-        Account as _, InputSource, WalletCommitmentTrees, WalletRead, WalletWrite,
+        WalletCommitmentTrees, WalletRead, WalletWrite,
         chain::{scan_cached_blocks, CommitmentTreeRoot},
         scanning::ScanPriority,
-        wallet::{
-            self, ConfirmationsPolicy,
-            input_selection::GreedyInputSelector,
-            propose_transfer, create_proposed_transactions,
-        },
+        wallet::ConfirmationsPolicy,
     },
-    fees::{
-        DustOutputPolicy, SplitPolicy, StandardFeeRule,
-        zip317::MultiOutputChangeStrategy,
-    },
+    fees::StandardFeeRule,
     proto::service::TreeState,
-    wallet::OvkPolicy,
-    zip321::{Payment, TransactionRequest},
 };
 use zcash_client_sqlite::{
     AccountUuid, FsBlockDb,
@@ -28,20 +17,19 @@ use zcash_client_sqlite::{
     chain::{BlockMeta, init::init_blockmeta_db},
     util::SystemClock,
 };
-use crate::wallet::keys::parse_account_uuid;
-use zcash_keys::keys::UnifiedSpendingKey;
 use zcash_primitives::block::BlockHash;
 use zcash_proofs::prover::LocalTxProver;
-use zcash_protocol::{
-    ShieldedProtocol,
-    consensus::{BlockHeight, Network},
-    memo::{Memo, MemoBytes},
-    value::Zatoshis,
-};
+use zcash_protocol::consensus::{BlockHeight, Network};
 
-type WalletDatabase = WalletDb<rusqlite::Connection, Network, SystemClock, OsRng>;
+use crate::wallet::keys::parse_account_uuid;
 
-fn open_wallet_db(db_path: &str, network: Network) -> Result<WalletDatabase, String> {
+mod send;
+
+pub(crate) use send::{estimate_fee, execute_proposal, propose_send};
+
+pub(super) type WalletDatabase = WalletDb<rusqlite::Connection, Network, SystemClock, OsRng>;
+
+pub(super) fn open_wallet_db(db_path: &str, network: Network) -> Result<WalletDatabase, String> {
     WalletDb::for_path(db_path, network, SystemClock, OsRng)
         .map_err(|e| format!("Failed to open wallet DB: {e}"))
 }
@@ -231,227 +219,34 @@ pub fn validate_address(address: &str) -> Result<String, String> {
 
 /// Propose a transfer. Returns (proposal_id, needs_sapling_params, fee_zatoshi).
 /// The proposal is stored internally and referenced by proposal_id for execute_proposal.
-pub fn propose_send(
-    db_path: &str, network: Network, account_uuid: &str,
-    to_address: &str, amount_zatoshi: u64, memo_str: Option<&str>,
-) -> Result<ProposalResult, String> {
-    use zcash_protocol::{PoolType, ShieldedProtocol as SP};
-
-    let mut db = open_wallet_db(db_path, network)?;
-    let account_id = parse_account_uuid(account_uuid)?;
-
-    let to: zcash_address::ZcashAddress = to_address.parse().map_err(|e| format!("Bad address: {e}"))?;
-    let value = Zatoshis::from_u64(amount_zatoshi).map_err(|_| "Bad amount")?;
-    let memo_bytes = match memo_str {
-        Some(m) => {
-            let bytes = MemoBytes::from(Memo::from_bytes(m.as_bytes()).map_err(|e| format!("Bad memo: {e}"))?);
-            Some(bytes)
-        }
-        None => None,
-    };
-
-    let (change_strategy, input_selector) = zip317_helper::<WalletDatabase>(None);
-    let payment = Payment::new(to, value, memo_bytes, None, None, vec![])
-        .ok_or("Cannot send memo to this address type")?;
-    let request = TransactionRequest::new(vec![payment]).map_err(|e| format!("{e:?}"))?;
-
-    let proposal = propose_transfer::<_, _, _, _, Infallible>(
-        &mut db, &network, account_id, &input_selector, &change_strategy,
-        request, ConfirmationsPolicy::default(),
-    ).map_err(|e| format!("Propose failed: {e}"))?;
-
-    let needs_sapling = proposal.steps().iter().any(|step| {
-        step.involves(PoolType::Shielded(SP::Sapling))
-    });
-
-    let fee: u64 = proposal.steps().iter()
-        .map(|step| u64::from(step.balance().fee_required()))
-        .sum();
-
-    // Store proposal for later execution
-    let mut store = PROPOSAL_STORE.lock().map_err(|e| format!("Lock error: {e}"))?;
-    let id = store.next_id;
-    store.next_id += 1;
-    store.proposals.insert(id, StoredProposal { proposal, network, account_id });
-
-    Ok(ProposalResult { proposal_id: id, needs_sapling_params: needs_sapling, fee_zatoshi: fee })
-}
-
-/// Estimate the fee for a transfer without storing the proposal.
-/// Used for validation only — does not consume resources in PROPOSAL_STORE.
-pub fn estimate_fee(
-    db_path: &str, network: Network, account_uuid: &str,
-    to_address: &str, amount_zatoshi: u64, memo_str: Option<&str>,
-) -> Result<u64, String> {
-    use zcash_protocol::ShieldedProtocol as SP;
-
-    let mut db = open_wallet_db(db_path, network)?;
-    let account_id = parse_account_uuid(account_uuid)?;
-
-    let to: zcash_address::ZcashAddress = to_address.parse().map_err(|e| format!("Bad address: {e}"))?;
-    let value = Zatoshis::from_u64(amount_zatoshi).map_err(|_| "Bad amount")?;
-    let memo_bytes = match memo_str {
-        Some(m) => {
-            let bytes = MemoBytes::from(Memo::from_bytes(m.as_bytes()).map_err(|e| format!("Bad memo: {e}"))?);
-            Some(bytes)
-        }
-        None => None,
-    };
-
-    let (change_strategy, input_selector) = zip317_helper::<WalletDatabase>(None);
-    let payment = Payment::new(to, value, memo_bytes, None, None, vec![])
-        .ok_or("Cannot send memo to this address type")?;
-    let request = TransactionRequest::new(vec![payment]).map_err(|e| format!("{e:?}"))?;
-
-    let proposal = propose_transfer::<_, _, _, _, Infallible>(
-        &mut db, &network, account_id, &input_selector, &change_strategy,
-        request, ConfirmationsPolicy::default(),
-    ).map_err(|e| format!("Propose failed: {e}"))?;
-
-    let fee: u64 = proposal.steps().iter()
-        .map(|step| u64::from(step.balance().fee_required()))
-        .sum();
-
-    Ok(fee)
-}
-
-/// Execute a previously proposed transfer, then broadcast to the network.
-/// Returns comma-separated txids on success.
-pub async fn execute_proposal(
-    db_path: &str,
-    lightwalletd_url: &str,
-    proposal_id: u64,
-    seed_bytes: &[u8],
-    spend_params_path: Option<&str>,
-    output_params_path: Option<&str>,
-) -> Result<String, String> {
-    let mut store = PROPOSAL_STORE.lock().map_err(|e| format!("Lock error: {e}"))?;
-    let stored = store.proposals.remove(&proposal_id)
-        .ok_or("Proposal not found (expired or already executed)")?;
-    let network = stored.network;
-    drop(store);
-
-    let mut db = open_wallet_db(db_path, network)?;
-    let account_id = stored.account_id;
-    let account = db.get_account(account_id).map_err(|e| format!("{e}"))?.ok_or("Account not found")?;
-
-    // Scope seed/USK so they are dropped before network I/O (broadcast)
-    let txids = {
-        let seed = SecretVec::new(seed_bytes.to_vec());
-        let zip32_index = account.source().key_derivation().ok_or("No key derivation")?.account_index();
-        let usk = UnifiedSpendingKey::from_seed(&network, seed.expose_secret(), zip32_index)
-            .map_err(|e| format!("USK derivation failed: {e:?}"))?;
-
-        match (spend_params_path, output_params_path) {
-            (Some(sp), Some(op)) if !sp.is_empty() && !op.is_empty() => {
-                let prover = LocalTxProver::new(
-                    std::path::Path::new(sp),
-                    std::path::Path::new(op),
-                );
-                create_proposed_transactions::<_, _, Infallible, _, Infallible, _>(
-                    &mut db, &network, &prover, &prover,
-                    &wallet::SpendingKeys::from_unified_spending_key(usk),
-                    OvkPolicy::Sender, &stored.proposal,
-                ).map_err(|e| format!("Create TX failed: {e}"))?
-            }
-            _ => {
-                let spend_prover = NoOpSpendProver;
-                let output_prover = NoOpOutputProver;
-                create_proposed_transactions::<_, _, Infallible, _, Infallible, _>(
-                    &mut db, &network, &spend_prover, &output_prover,
-                    &wallet::SpendingKeys::from_unified_spending_key(usk),
-                    OvkPolicy::Sender, &stored.proposal,
-                ).map_err(|e| format!("Create TX failed: {e}"))?
-            }
-        }
-        // seed + usk dropped here, before broadcast
-    };
-
-    // Connect to lightwalletd once for all broadcasts.
-    let (mut client, _tor_guard) =
-        crate::wallet::sync_engine::open_lwd_channel(lightwalletd_url)
-            .await
-            .map_err(|e| e.to_string())?;
-
-    // Broadcast each transaction
-    let txid_strings: Vec<String> = txids.iter().map(|id| format!("{id}")).collect();
-
-    let read_conn = rusqlite::Connection::open_with_flags(
-        db_path, rusqlite::OpenFlags::SQLITE_OPEN_READ_ONLY,
-    ).map_err(|e| format!("Failed to open DB for broadcast: {e}"))?;
-
-    let mut broadcast_ok: Vec<String> = Vec::new();
-    for txid in &txids {
-        let raw_tx = read_conn
-            .query_row(
-                "SELECT raw FROM transactions WHERE txid = ?1",
-                rusqlite::params![txid.as_ref()],
-                |row| row.get::<_, Vec<u8>>(0),
-            )
-            .map_err(|e| format!("Failed to get raw tx for {txid}: {e}"))?;
-
-        match broadcast_raw_transaction(&mut client, &raw_tx).await {
-            Ok(()) => {
-                broadcast_ok.push(format!("{txid}"));
-                log::info!("send: broadcast {txid} ({} bytes)", raw_tx.len());
-            }
-            Err(e) => {
-                return Err(format!(
-                    "Broadcast failed after {}/{} txs sent ({}). Error: {e}",
-                    broadcast_ok.len(), txids.len(), broadcast_ok.join(",")
-                ));
-            }
-        }
-    }
-
-    Ok(txid_strings.join(","))
-}
-
-/// Broadcast a raw transaction using an existing gRPC client.
-async fn broadcast_raw_transaction(
-    client: &mut zcash_client_backend::proto::service::compact_tx_streamer_client::CompactTxStreamerClient<tonic::transport::Channel>,
-    raw_tx: &[u8],
-) -> Result<(), String> {
-    use zcash_client_backend::proto::service::RawTransaction;
-
-    let resp = client
-        .send_transaction(RawTransaction {
-            data: raw_tx.to_vec(),
-            height: 0,
-        })
-        .await
-        .map_err(|e| format!("SendTransaction gRPC failed: {e}"))?
-        .into_inner();
-
-    if resp.error_code != 0 {
-        return Err(format!("Broadcast rejected: {} (code {})", resp.error_message, resp.error_code));
-    }
-
-    Ok(())
-}
-
-pub(crate) struct ProposalResult {
-    pub proposal_id: u64,
-    pub needs_sapling_params: bool,
-    pub fee_zatoshi: u64,
-}
-
-// In-memory proposal store (proposals are short-lived, between propose and execute)
+// In-memory proposal store (proposals are short-lived, between
+// propose and execute). Kept in `sync/mod.rs` because it is shared
+// between the software send flow (`send::execute_proposal`) and the
+// hardware PCZT pipeline (`pczt::create_pczt_from_proposal`); placing
+// it in either submodule would create a cross-submodule dependency.
 use std::collections::HashMap;
 use std::sync::Mutex;
 
-struct StoredProposal {
-    proposal: zcash_client_backend::proposal::Proposal<StandardFeeRule, zcash_client_sqlite::ReceivedNoteId>,
-    network: Network,
-    account_id: AccountUuid,
+pub(super) struct StoredProposal {
+    pub proposal: zcash_client_backend::proposal::Proposal<
+        StandardFeeRule,
+        zcash_client_sqlite::ReceivedNoteId,
+    >,
+    pub network: Network,
+    pub account_id: AccountUuid,
 }
 
-static PROPOSAL_STORE: std::sync::LazyLock<Mutex<ProposalStore>> =
-    std::sync::LazyLock::new(|| Mutex::new(ProposalStore { proposals: HashMap::new(), next_id: 1 }));
+pub(super) static PROPOSAL_STORE: std::sync::LazyLock<Mutex<ProposalStore>> =
+    std::sync::LazyLock::new(|| {
+        Mutex::new(ProposalStore {
+            proposals: HashMap::new(),
+            next_id: 1,
+        })
+    });
 
-struct ProposalStore {
-    proposals: HashMap<u64, StoredProposal>,
-    next_id: u64,
+pub(super) struct ProposalStore {
+    pub proposals: HashMap<u64, StoredProposal>,
+    pub next_id: u64,
 }
 
 // ======================== PCZT (Hardware Wallet) ========================
@@ -914,85 +709,6 @@ pub fn get_transaction_history(
         .map_err(|e| format!("Row error: {e}"))
 }
 
-// ======================== No-op Sapling Provers ========================
-// Used for Orchard-only transactions where Sapling params are not available.
-// The prover methods will never be called for Orchard-only TXs.
-
-use sapling_crypto::{
-    bundle::GrothProofBytes,
-    circuit,
-    keys::EphemeralSecretKey,
-    prover::{OutputProver, SpendProver},
-    value::{NoteValue, ValueCommitTrapdoor},
-    Diversifier, MerklePath, PaymentAddress, ProofGenerationKey, Rseed,
-};
-
-const GROTH_PROOF_SIZE: usize = 192;
-
-struct NoOpSpendProver;
-
-impl SpendProver for NoOpSpendProver {
-    type Proof = GrothProofBytes;
-
-    fn prepare_circuit(
-        _proof_generation_key: ProofGenerationKey,
-        _diversifier: Diversifier,
-        _rseed: Rseed,
-        _value: NoteValue,
-        _alpha: jubjub::Fr,
-        _rcv: ValueCommitTrapdoor,
-        _anchor: bls12_381::Scalar,
-        _merkle_path: MerklePath,
-    ) -> Option<circuit::Spend> {
-        log::error!("NoOpSpendProver::prepare_circuit called — proposal contains unexpected Sapling spend");
-        None
-    }
-
-    fn create_proof<R: rand_core::RngCore>(
-        &self, _circuit: circuit::Spend, _rng: &mut R,
-    ) -> Self::Proof {
-        log::error!("NoOpSpendProver::create_proof called — should never happen");
-        [0u8; GROTH_PROOF_SIZE]
-    }
-
-    fn encode_proof(_proof: Self::Proof) -> GrothProofBytes {
-        [0u8; GROTH_PROOF_SIZE]
-    }
-}
-
-struct NoOpOutputProver;
-
-impl OutputProver for NoOpOutputProver {
-    type Proof = GrothProofBytes;
-
-    fn prepare_circuit(
-        _esk: &EphemeralSecretKey,
-        _payment_address: PaymentAddress,
-        _rcm: jubjub::Fr,
-        _value: NoteValue,
-        _rcv: ValueCommitTrapdoor,
-    ) -> circuit::Output {
-        log::error!("NoOpOutputProver::prepare_circuit called — proposal contains unexpected Sapling output");
-        circuit::Output {
-            value_commitment_opening: None,
-            payment_address: None,
-            commitment_randomness: None,
-            esk: None,
-        }
-    }
-
-    fn create_proof<R: rand_core::RngCore>(
-        &self, _circuit: circuit::Output, _rng: &mut R,
-    ) -> Self::Proof {
-        log::error!("NoOpOutputProver::create_proof called — should never happen");
-        [0u8; GROTH_PROOF_SIZE]
-    }
-
-    fn encode_proof(_proof: Self::Proof) -> GrothProofBytes {
-        [0u8; GROTH_PROOF_SIZE]
-    }
-}
-
 // ======================== Pending TX Tracking ========================
 
 pub(crate) struct PendingTxInfo {
@@ -1066,19 +782,6 @@ pub async fn check_tx_mined(lightwalletd_url: &str, txid_bytes: &[u8]) -> i64 {
 
 pub fn get_blocks_dir(cache_path: &str) -> String {
     format!("{cache_path}/blocks")
-}
-
-fn zip317_helper<DbT: InputSource>(
-    change_memo: Option<MemoBytes>,
-) -> (MultiOutputChangeStrategy<StandardFeeRule, DbT>, GreedyInputSelector<DbT>) {
-    (
-        MultiOutputChangeStrategy::new(
-            StandardFeeRule::Zip317, change_memo, ShieldedProtocol::Orchard,
-            DustOutputPolicy::default(),
-            SplitPolicy::with_min_output_value(NonZeroUsize::new(4).unwrap(), Zatoshis::const_from_u64(1000_0000)),
-        ),
-        GreedyInputSelector::new(),
-    )
 }
 
 #[cfg(test)]
